+++ activities/ColorMyWorld.activity/lib/map.js
@@ -29,7 +29,7 @@ define(["activity/ol","print","util","colormyworld","humane","flag","l10n"],
 					if (!me.tooltipDisplay || target_name!=me.tooltipDisplay) {
 						me.tooltipDisplay=target_name;
 						humane.timeout=1000;
-						humane.log(flag[`${target_name.replace(/ /g,'_')}`]+" "+ l10n.get(target_name.replace(/ /g,'_')).replace(/_/g,' '));
+                                                humane.log("<img src='./flags/"+flag[`${target_name.replace(/ /g,'_')}`]+".svg' style='width: auto; height: 1em;'>&nbsp;&nbsp;"+l10n.get(target_name.replace(/ /g,'_')).replace(/_/g,' '));
 						setTimeout(function() {
 							me.tooltipDisplay=null;
 						}, humane.timeout);